Add execution context stack for accessing caller information
This commit adds an execution context stack to the tracer.
Every time a closure, builtin or special is entered, a
context is pushed on the stack and on exiting from the function
the context is popped.

This provides access to the entire call chain. This can be used
to get information such as the caller of the currently entered
function or the parent environment.

eval expression analysis has also been updated to utilize the
stack and get the caller of the eval function.